adding new enemy and tower need to add cards for new tower MySlider
fixed red errors
new card idea
new locked card idea

constant WORK

need add good-looking
button for every thing in
asprite
